+++ src/config/tabs/KeyboardConfigTab.cpp
@@ -45,6 +45,7 @@ static std::array const SHORTCUT_NAMES = {
     QT_TR_NOOP("Transpose, increase octave"),
     QT_TR_NOOP("Transpose..."),
     QT_TR_NOOP("Reverse"),
+    QT_TR_NOOP("Replace instrument"),
     QT_TR_NOOP("Play"),
     QT_TR_NOOP("Play from start"),
     QT_TR_NOOP("Play from cursor"),